Change of mind for supporting MIRAGE
Instead of supporting separate meta and data arrays in `CacheArray',
it is better to have a separate vector of `CacheArray' for metadata and one dedicated CacheArrays for data in `CacheBase'.
When the data array is separated, it need to keep a small metadata as well, which becomes a full `CacheArray'.
In this way, there is no need to support locating data using different set and way in `CacheArray'.