bugfix: fix a delicate concurrency bug
Making the pool state as anatomic variable is not safe enough to protect
the buffer pool. When returning a buffer, the buffer state is updated
atomically. Another thread may waiting on this event and immediately
allocate this buffer before the actual pointer is written back to the
pool. As a result, the waiting thread may write on a stale buffer and
lead to silent data corruption later on.

The solution is reverting back to the mutex lock which guarantee both the
pool and the state are updated as a bundle.