Add a magic interface for PFC monitors
This magic interface provide a special channel for cache to talk
with an underliying PFC monitor with non-uniform requirement.
For example, a cache capable of relocating cache blocks may need
to count the number of relocations in a period of time. This info
cannot be collected using the typical PFC interface. With this
magic interface, the relocation method can directly call a monitor
using a predefined magic id, to record such events. When none of
such monitor is available/enabled/attached, there is barely any
side-effect expcet for a fruitless seasrch of all attached monitors,
which normally are with a tiny number.